content: move new "Artists - by Name" implementation to "by Group"
