+++ src/strings-default.json
@@ -368,8 +368,6 @@
   "listingPage.listArtists.byName.title": "Artists - by Name",
   "listingPage.listArtists.byName.title.short": "...by Name",
   "listingPage.listArtists.byName.item": "{ARTIST} ({CONTRIBUTIONS})",
-  "listingPage.listArtists.byName.chunk.title": "Contributed to {GROUP}:",
-  "listingPage.listArtists.byName.chunk.item": "{ARTIST} ({CONTRIBUTIONS})",
   "listingPage.listArtists.byContribs.title": "Artists - by Contributions",
   "listingPage.listArtists.byContribs.title.short": "...by Contributions",
   "listingPage.listArtists.byContribs.item": "{ARTIST} ({CONTRIBUTIONS})",
@@ -379,6 +377,11 @@
   "listingPage.listArtists.byDuration.title": "Artists - by Duration",
   "listingPage.listArtists.byDuration.title.short": "...by Duration",
   "listingPage.listArtists.byDuration.item": "{ARTIST} ({DURATION})",
+  "listingPage.listArtists.byGroup.title": "Artists - by Group",
+  "listingPage.listArtists.byGroup.title.short": "...by Group",
+  "listingPage.listArtists.byGroup.item": "{ARTIST} ({CONTRIBUTIONS})",
+  "listingPage.listArtists.byGroup.chunk.title": "Contributed to {GROUP}:",
+  "listingPage.listArtists.byGroup.chunk.item": "{ARTIST} ({CONTRIBUTIONS})",
   "listingPage.listArtists.byLatest.title": "Artists - by Latest Contribution",
   "listingPage.listArtists.byLatest.title.short": "...by Latest Contribution",
   "listingPage.listArtists.byLatest.chunk.title.album": "{ALBUM} ({DATE})",
